stream chat completions for oracle
