+++ src/providers/oracle/index.ts
@@ -3,6 +3,7 @@ import OracleAPIConfig from './api';
 import {
   OracleChatCompleteConfig,
   OracleChatCompleteResponseTransform,
+  OracleChatCompleteStreamChunkTransform,
 } from './chatComplete';
 
 const OracleConfig: ProviderConfigs = {
@@ -10,6 +11,7 @@ const OracleConfig: ProviderConfigs = {
   api: OracleAPIConfig,
   responseTransforms: {
     chatComplete: OracleChatCompleteResponseTransform,
+    'stream-chatComplete': OracleChatCompleteStreamChunkTransform,
   },
 };
